predictive-maintenance/agents/tool-maintenance/pyproject.toml
[project]
name = "tool"
version = "0.1.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastmcp>=2.10.6",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
]